requirements.txt
cltl.backend[impl,host,service]
cltl.asr[impl,service]
cltl.vad[impl,service]
cltl.chat-ui
cltl.eliza

flask
werkzeug
